feat(drawsystem): separate DrawMachine and DrawResult
on constructing DrawMachine methods
add new DrawResult class into drawsystem to separate roles clearly

BREAKING CHANGE: change type of mainNumbers in WinningNumbers
 - from TreeSet<Integer> to List<Integer>
