+++ src/main/java/lotto/committee/WinningNumbers.java
@@ -1,13 +1,15 @@
 package lotto.committee;
 
+import java.util.List;
 import java.util.TreeSet;
+import lotto.drawsystem.DrawMachine;
 
 public class WinningNumbers {
 
-    private final TreeSet<Integer> mainNumbers;
+    private final List<Integer> mainNumbers;
     private final Integer bonusNumber;
 
-    private WinningNumbers(TreeSet<Integer> mainNumbers, Integer bonusNumber) {
+    private WinningNumbers(List<Integer> mainNumbers, Integer bonusNumber) {
         this.mainNumbers = mainNumbers;
         this.bonusNumber = bonusNumber;
     }
@@ -16,11 +18,17 @@ public class WinningNumbers {
         return new WinningNumbers(mainNumbers, bonusNumber);
     }
 
-    public TreeSet<Integer> getMainNumbers() {
+    public List<Integer> getMainNumbers() {
         return mainNumbers;
     }
 
     public Integer getBonusNumber() {
         return bonusNumber;
+    }
+
+    private WinningNumbers runWinningNumbers() {
+        DrawMachine drawMachine = new DrawMachine();
+         = drawMachine.runSingleDraw();
+
     }
 }
