+++ src/main/java/lotto/drawsystem/DrawMachine.java
@@ -1,13 +1,23 @@
 package lotto.drawsystem;
 
-import java.util.TreeSet;
-
 public class DrawMachine {
 
-    private Integer drawnMainNumber;
-    private Integer drawnBonusNumber;
-    private final TreeSet<Integer> mainNumbers;
-    private final Integer bonusNumber;
+    public DrawMachine runDraws(Integer amount) {
+        DrawMachine drawMachine = new DrawMachine();
+        for(int drawCount = 0; drawCount < amount; amount++) {
+            runSingleDraw();
+        }
+        return drawMachine;
+    }
+
+
+    public DrawMachine runSingleDraw() {
+        Integer drawnMainNumber;
+        Integer drawnBonusNumber;
+
+
+        return DrawMachine();
+    }
