+++ src/main/java/lotto/drawsystem/DrawResult.java
@@ -1,0 +1,5 @@
+package lotto.drawsystem;
+
+public class DrawResult {
+
+}
